requirements.txt
schedule
google-generativeai
requests
numpy
